Moved the CPU capability logic inside fipsmodule (#330)
* Turn off FIPS static build in CI
* Rename the env var to OPENSSL_armcap to be consistent with x86
* Removed get_armcap_for_test function
* Removed OPENSSL_ia32cap_get() function
* Add target arch for all_tests.go
* Move the call to OPENSSL_cpuid_setup to bcm.c in FIPS mode
