+++ crypto/fipsmodule/cpucap/cpu-aarch64-fuchsia.c
@@ -23,8 +23,6 @@
 
 #include <openssl/arm_arch.h>
 
-#include "internal.h"
-
 extern uint32_t OPENSSL_armcap_P;
 
 void OPENSSL_cpuid_setup(void) {
