+++ crypto/fipsmodule/rand/internal.h
@@ -119,14 +119,14 @@ OPENSSL_EXPORT void CTR_DRBG_clear(CTR_DRBG_STATE *drbg);
 #if defined(OPENSSL_X86_64) && !defined(OPENSSL_NO_ASM)
 
 OPENSSL_INLINE int have_rdrand(void) {
-  return (OPENSSL_ia32cap_get()[1] & (1u << 30)) != 0;
+  return (OPENSSL_ia32cap_P[1] & (1u << 30)) != 0;
 }
 
 // have_fast_rdrand returns true if RDRAND is supported and it's reasonably
 // fast. Concretely the latter is defined by whether the chip is Intel (fast) or
 // not (assumed slow).
 OPENSSL_INLINE int have_fast_rdrand(void) {
-  const uint32_t *const ia32cap = OPENSSL_ia32cap_get();
+  const uint32_t *const ia32cap = OPENSSL_ia32cap_P;
   return (ia32cap[1] & (1u << 30)) && (ia32cap[0] & (1u << 30));
 }
