+++ crypto/test/gtest_main.cc
@@ -24,12 +24,6 @@
 #include "gtest_main.h"
 #include "../internal.h"
 
-#if (defined(OPENSSL_ARM) || defined(OPENSSL_AARCH64)) &&       \
-    !defined(OPENSSL_STATIC_ARMCAP)
-#include <openssl/arm_arch.h>
-#define TEST_ARM_CPUS
-#endif
-
 
 int main(int argc, char **argv) {
   testing::InitGoogleTest(&argc, argv);
@@ -42,33 +36,6 @@ int main(int argc, char **argv) {
       RAND_enable_fork_unsafe_buffering(-1);
     }
 #endif
-
-#if defined(TEST_ARM_CPUS)
-    if (strncmp(argv[i], "--cpu=", 6) == 0) {
-      const char *cpu = argv[i] + 6;
-      uint32_t armcap;
-      if (strcmp(cpu, "none") == 0) {
-        armcap = 0;
-      } else if (strcmp(cpu, "neon") == 0) {
-        armcap = ARMV7_NEON;
-      } else if (strcmp(cpu, "crypto") == 0) {
-        armcap = ARMV7_NEON | ARMV8_AES | ARMV8_SHA1 | ARMV8_SHA256 | ARMV8_PMULL;
-      } else {
-        fprintf(stderr, "Unknown CPU: %s\n", cpu);
-        exit(1);
-      }
-
-      uint32_t *armcap_ptr = OPENSSL_get_armcap_pointer_for_test();
-      if ((armcap & *armcap_ptr) != armcap) {
-        fprintf(stderr,
-                "Host CPU does not support features for testing CPU '%s'.\n",
-                cpu);
-        exit(89);
-      }
-      printf("Simulating CPU '%s'\n", cpu);
-      *armcap_ptr = armcap;
-    }
-#endif  // TEST_ARM_CPUS
 
     if (strcmp(argv[i], "--no_unwind_tests") == 0) {
       unwind_tests = false;
