+++ tests/ci/run_fips_tests.sh
@@ -4,37 +4,23 @@
 
 source tests/ci/common_posix_setup.sh
 
+echo "Testing AWS-LC in FIPS Release mode."
+fips_build_and_test -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED_LIBS=1
+
+# The AL2 version of Clang does not have all of the required artifacts for address sanitizer, see P45594051
+if [[ "${AWSLC_NO_ASM_FIPS}" == "1" ]]; then
+  echo "Building with Clang and testing AWS-LC in FIPS RelWithDebInfo mode with address sanitizer."
+  fips_build_and_test -DASAN=1 -DCMAKE_BUILD_TYPE=RelWithDebInfo -DBUILD_SHARED_LIBS=1
+else
+  echo "Not building with Clang and testing AWS-LC in FIPS RelWithDebInfo mode."
+  fips_build_and_test -DCMAKE_BUILD_TYPE=RelWithDebInfo -DBUILD_SHARED_LIBS=1
+fi
+
 if [[ ("$(uname -p)" == 'x86_64'*) ]]; then
   echo "Testing AWS-LC in FIPS Release and DISABLE_GETAUXVAL mode."
   fips_build_and_test -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED_LIBS=1 -DDISABLE_GETAUXVAL=1
 fi
 
-echo "Testing AWS-LC in FIPS Debug mode."
-fips_build_and_test
-
-# TODO(shang): investigate delocate transform errors when building static awslc with gcc fips release on aarch.
-if [[ ("${CC}" == 'clang'*) || ("$(uname -p)" == 'x86_64'*) ]]; then
-  echo "Testing AWS-LC in FIPS Release mode."
-  fips_build_and_test -DCMAKE_BUILD_TYPE=Release
-
-  echo "Testing AWS-LC in FIPS RelWithDebInfo mode."
-  fips_build_and_test -DCMAKE_BUILD_TYPE=RelWithDebInfo
-fi
-
-echo "Testing shared AWS-LC in FIPS Debug mode."
-fips_build_and_test -DBUILD_SHARED_LIBS=1
-
-echo "Testing shared AWS-LC in FIPS Release mode in a different folder."
+echo "Testing shared AWS-LC in FIPS Debug mode in a different folder."
 BUILD_ROOT=$(realpath "${SRC_ROOT}/../aws-lc-external_build")
-fips_build_and_test -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED_LIBS=1
-
-echo "Testing shared AWS-LC in FIPS RelWithDebInfo mode."
-fips_build_and_test -DCMAKE_BUILD_TYPE=RelWithDebInfo -DBUILD_SHARED_LIBS=1
-
-if [[  "${AWSLC_NO_ASM_FIPS}" == "1" ]]; then
-  # This dimension corresponds to boringssl CI 'linux_fips_noasm_asan'.
-  # https://logs.chromium.org/logs/boringssl/buildbucket/cr-buildbucket.appspot.com/8852496158370398336/+/steps/cmake/0/logs/execution_details/0
-  echo "Testing AWS-LC in FIPS OPENSSL_NO_ASM Release mode."
-  fips_build_and_test -DASAN=1 -DOPENSSL_NO_ASM=1 -DCMAKE_BUILD_TYPE=Release
-  # TODO(shang): add ASAN build with RelWithDebInfo if final FIPS build uses the option.
-fi
+fips_build_and_test -DCMAKE_BUILD_TYPE=Debug -DBUILD_SHARED_LIBS=1
